ai_matching/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
numpy==1.26.0
openai==1.3.7
scikit-learn==1.3.2
pandas==2.1.1
sentence-transformers==2.2.2
transformers==4.34.0
torch==2.0.1 